Rename event to events
