+++ src/main.rb
@@ -10,7 +10,6 @@ require 'miner/worker'
 require 'miner/coins'
 require 'miner/miners'
 require 'api/mph'
-require 'miner/event'
 require 'miner/triggers'
 require 'miner/actions'
 require 'util/application'
