+++ src/miner/events.rb
